+++ main.py
@@ -1,5 +1,5 @@
 import asyncio
-from controller.run_pipeline import run
+from controller.crawler_pipeline import run
 
 if __name__ == "__main__":
     asyncio.run(run())
